# docs/production/openapi.json
{
  "info": {
    "description": "Generated from the route table (server/build/routeTable.ts) and the request-contract registry (server/build/apiSurface.ts) by scripts/generate-openapi.ts. Request bodies are the zod schemas the handlers parse with, emitted by zod. Success response bodies are not yet described; x-coverage says how much of the surface each part of this document covers.",
    "title": "Abedin Growth AI API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/contacts/{survivorId}/merge": {
      "post": {
        "operationId": "post_api_contacts_survivorid_merge",
        "parameters": [
          {
            "in": "path",
            "name": "survivorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "ATTRIBUTION_REQUIRED",
                            "AUTH_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_UNAVAILABLE",
                            "CONFIGURATION_ERROR",
                            "CONTACT_EXISTS",
                            "ILLEGAL_TRANSITION",
                            "INTERNAL_ERROR",
                            "MERGE_REFUSED",
                            "MODEL_OUTPUT_INVALID",
                            "NOT_FOUND",
                            "NOT_IMPLEMENTED",
                            "POLICY_BLOCKED",
                            "PROVIDER_RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "STORE_UNAVAILABLE",
                            "TENANT_AMBIGUOUS",
                            "TENANT_FORBIDDEN",
                            "TENANT_INVALID",
                            "TENANT_REVOCATION_UNVERIFIABLE",
                            "TENANT_SUSPENDED",
                            "TENANT_UNRESOLVED",
                            "TERMINAL_STATE",
                            "TOO_MANY_REFERENCES",
                            "UNKNOWN_CURRENT_STATE",
                            "UNKNOWN_TARGET_STATE",
                            "VALIDATION_ERROR",
                            "VERSION_CONFLICT",
                            "VERSION_REQUIRED",
                            "WEBHOOK_VERIFICATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Any error, in the envelope server/lib/errors.ts builds; the code is one of the taxonomy."
          }
        },
        "x-body": "no-schema",
        "x-body-note": "The handler reads req.body and no schema describes it, so this document cannot either; some of these check fields by hand. Listed by name in openapi.invariant.test.ts; the list may shrink and not grow.",
        "x-source": "server/routes/contacts.routes.ts:179"
      }
    },
...
    "/api/investors": {
      "get": {
        "operationId": "get_api_investors",
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "ATTRIBUTION_REQUIRED",
                            "AUTH_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_UNAVAILABLE",
                            "CONFIGURATION_ERROR",
                            "CONTACT_EXISTS",
                            "ILLEGAL_TRANSITION",
                            "INTERNAL_ERROR",
                            "MERGE_REFUSED",
                            "MODEL_OUTPUT_INVALID",
                            "NOT_FOUND",
                            "NOT_IMPLEMENTED",
                            "POLICY_BLOCKED",
                            "PROVIDER_RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "STORE_UNAVAILABLE",
                            "TENANT_AMBIGUOUS",
                            "TENANT_FORBIDDEN",
                            "TENANT_INVALID",
                            "TENANT_REVOCATION_UNVERIFIABLE",
                            "TENANT_SUSPENDED",
                            "TENANT_UNRESOLVED",
                            "TERMINAL_STATE",
                            "TOO_MANY_REFERENCES",
                            "UNKNOWN_CURRENT_STATE",
                            "UNKNOWN_TARGET_STATE",
                            "VALIDATION_ERROR",
                            "VERSION_CONFLICT",
                            "VERSION_REQUIRED",
                            "WEBHOOK_VERIFICATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Any error, in the envelope server/lib/errors.ts builds; the code is one of the taxonomy."
          }
        },
        "x-body": "none",
        "x-source": "server/routes/contacts.routes.ts:349"
      },
      "post": {
        "operationId": "post_api_investors",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "companyName": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "companyWebsite": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "country": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "email": {
                    "maxLength": 320,
                    "type": "string"
                  },
                  "employeeCount": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "firstName": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "industry": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "lastName": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "linkedinUrl": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "notes": {
                    "maxLength": 10000,
                    "type": "string"
                  },
                  "phone": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "title": {
                    "maxLength": 500,
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "ATTRIBUTION_REQUIRED",
                            "AUTH_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_UNAVAILABLE",
                            "CONFIGURATION_ERROR",
                            "CONTACT_EXISTS",
                            "ILLEGAL_TRANSITION",
                            "INTERNAL_ERROR",
                            "MERGE_REFUSED",
                            "MODEL_OUTPUT_INVALID",
                            "NOT_FOUND",
                            "NOT_IMPLEMENTED",
                            "POLICY_BLOCKED",
                            "PROVIDER_RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "STORE_UNAVAILABLE",
                            "TENANT_AMBIGUOUS",
                            "TENANT_FORBIDDEN",
                            "TENANT_INVALID",
                            "TENANT_REVOCATION_UNVERIFIABLE",
                            "TENANT_SUSPENDED",
                            "TENANT_UNRESOLVED",
                            "TERMINAL_STATE",
                            "TOO_MANY_REFERENCES",
                            "UNKNOWN_CURRENT_STATE",
                            "UNKNOWN_TARGET_STATE",
                            "VALIDATION_ERROR",
                            "VERSION_CONFLICT",
                            "VERSION_REQUIRED",
                            "WEBHOOK_VERIFICATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Any error, in the envelope server/lib/errors.ts builds; the code is one of the taxonomy."
          }
        },
        "x-body": "contract",
        "x-source": "server/routes/contacts.routes.ts:358"
      }
    },
...
    "/api/leads/{id}/email": {
      "post": {
        "operationId": "post_api_leads_id_email",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "ATTRIBUTION_REQUIRED",
                            "AUTH_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_UNAVAILABLE",
                            "CONFIGURATION_ERROR",
                            "CONTACT_EXISTS",
                            "ILLEGAL_TRANSITION",
                            "INTERNAL_ERROR",
                            "MERGE_REFUSED",
                            "MODEL_OUTPUT_INVALID",
                            "NOT_FOUND",
                            "NOT_IMPLEMENTED",
                            "POLICY_BLOCKED",
                            "PROVIDER_RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "STORE_UNAVAILABLE",
                            "TENANT_AMBIGUOUS",
                            "TENANT_FORBIDDEN",
                            "TENANT_INVALID",
                            "TENANT_REVOCATION_UNVERIFIABLE",
                            "TENANT_SUSPENDED",
                            "TENANT_UNRESOLVED",
                            "TERMINAL_STATE",
                            "TOO_MANY_REFERENCES",
                            "UNKNOWN_CURRENT_STATE",
                            "UNKNOWN_TARGET_STATE",
                            "VALIDATION_ERROR",
                            "VERSION_CONFLICT",
                            "VERSION_REQUIRED",
                            "WEBHOOK_VERIFICATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Any error, in the envelope server/lib/errors.ts builds; the code is one of the taxonomy."
          }
        },
        "x-body": "none",
        "x-source": "server/routes/contacts.routes.ts:333"
      }
    },
...
  },
...
}